Update FASTBuild to use VS2019 16.7.0 (14.27.29110)
 - fix various newly surfaced compiler warnings

+++ Code/Core/Process/Process.cpp
@@ -620,7 +620,7 @@ bool Process::ReadAllData( AString & outMem,
                 if ( result == WAIT_TIMEOUT )
                 {
                     // Check if timeout is hit
-                    if ( ( timeOutMS > 0 ) && ( t.GetElapsedMS() >= timeOutMS ) )
+                    if ( ( timeOutMS > 0 ) && ( t.GetElapsedMS() >= (float)timeOutMS ) )
                     {
                         Terminate();
                         return false; // Timed out
